pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
line-bot-sdk = "^3.5.1"
jwcrypto = "^1.5.0"
pyjwt = "^2.8.0"

[tool.poetry.dev-dependencies]
ruff = "^0.1.8"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]

[tool.pytest.ini_options]
testpaths = ["tests"]
